Add get_privilege_context_revision
Allow the engines to fetch the privilege context revision so
they may optimize privilege checks as running the privilege
check require the ScopeID and CollectionID to be present.

Looking up the ScopeID may be costly, and unless the privilege
context has changed it'll always return the same value for
the same input.

Change-Id: Ifcbc3448f442a30e31cdc4db038afc238a419328
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/123506
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>