MB-38197: On-disk #items incorrect after SyncWrite + Delete
If an item is added via a SyncWrite, but is deleted (in-memory)
_before_ the Commit of the SyncWrite is persisted to disk, then
onDiskTotalItems is incorrectly not incremented. This results in
under-counting of how many items exist in the VBucket for Full
Eviction buckets.

The bug is caused by an incorrect assumption in the Set
PersistenceCallback - that onDiskTotalItems should only be incremented
if a StoredValue was found in the HashTable.

This was a correct assumption prior to SyncWrites - a dirty, deleted
StoredValue would never be completely removed from the HashTable - at
most it could be "UnDeleted" and replaced with a non-Deleted
StoredValue.

However, SyncWrites can replace a dirty, deleted SV with a Pending
StoredValue, and given Pending SVs are not included in
onDiskTotalItems then the increment was skipped.

Change-Id: I30307b39c4c95aca93fac1a447ae99aa6f9ae765
Reviewed-on: http://review.couchbase.org/123523
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>