MB-37402: Use SeqnoAdvanced to ensure a consumer is aware of snapShotEnd
Use SeqnoAdvanced to ensure that a consumer is told that the there are
no more mutations to be sent by the producer. This is done by sending a
SeqnoAdvanced op which contains a seqno of the snapEnd to signify the
there are no more DcpOps to be sent in this snapshot.

For example:
1) A backfill snapshot for a vbucket contains no mutations for
collection A, but collection B has many mutations meaning the vbucket
seqno X > 2. But a backfill stream for collection A would be told about
the DcpSnapShotMarker and DcpSystemEvent but would then wait for other
mutations as the SnapEnd = X, which will have all been filtered out.
Thus adding a SeqnoAdvanced allows us to inform the consumer the
snapshot had come to the end.

2) The last item in the Snapshot isn't visible to the consumer. This can
occur if the consumer supports collections but does not support
sync writes. As the snapEnd would be sent as the vbucket high seqno but
the last item in the snapshot could be prepare or an abort. Thus, we
need to inform that snapshot has come to an end using the SeqnoAdvanced
Dcp opcode.

Change-Id: I37c06d7760064c84fa22dcfe0a4473f19006d02e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124097
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>