MB-38270: ns_server don't synchronize file updates
When ns_server wants to update one of the configuration files it'll
write the content into a new file, deletes the old and renames the
new file (this works around the problem on windows that you can't
write to a file someone have open). It does however introduce a
race with memcached that we might try to open the file in the
short period between the deletion and the rename.

To work around this I've added retry logic in cb::io::loadFile
(http://review.couchbase.org/#/c/123582/).

Accessing these files was written under the assumption that
ns_server wouldn't touch the files:

* in the time from memcached was started until it could log in
* in the time from it told memcached to reload the file until
  it received the response from memcached that it had read it

Change-Id: If467f8211e95fe0480a912bc1dd821a9b888d2bf
Reviewed-on: http://review.couchbase.org/123655
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>