MB-37233: Allow to track any metric in EP benchmarks
We this patch, our EP benchmark scripts for Commit Validation require
the user to specify (via cbnt_test_list.yml) the metrics that have to be
tracked.

The user must provide a comma-separated list of metrics via the new
'--cbnt_metrics' command-line argument. Any GBench or user-defined
metric is accepted.

Some clean up too: specifying the time-format was not used, but now it
can be removed as a metric could be anything (eg, real-time, mem-usage,
etc..).

As a last relevant change, the new metric selection is used to track the
proper quantity for the benchmarks that we are currently running as part
of CV. In general that means tracking:
 - cpu-time for memory-only benchs
 - real-time for IO benchs

Change-Id: Ifc5f3395ee359154855bd09fdc2cd274664191fa
Reviewed-on: http://review.couchbase.org/123722
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>