MB-37789: Make Rollback use correct read snapshot
Rollback was passing to the callback incorrect snapshots which were
used as part of the Get function to retrieve the meta data and value.
The Get function call is not even required since the EPDiskRollbackCB
does all the necessary work and we are pointing at the target rollback
point. We just need to pass in the key and meta data.

Change-Id: Icb244c6dd53ce143ae7febb98cb618783b26e716
Reviewed-on: http://review.couchbase.org/123609
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>