MB-32766: Never purge the default collection tombstone
Collection life time 101:

a) create a collection => we store a system event
  - "create of collection x @ seqno y"
b) drop a collection => we delete that system event (becomes a tombstone)
c) 1 hour 3 days later we purge the tombstone

However the default collection because we 'pretend' it always existed
we never write the create event (a), the creation is assumed. Thus
we can never safely purge the tombstone otherwise a client could
assume it exists when it does not.

To keep the tombstone, we set the 'delete' time to be ~0 so that normal
purging can never remove it.

Change-Id: I1b225d6886f0b30072252e515f342ee66ee7957a
Reviewed-on: http://review.couchbase.org/124109
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>