MB-38389: completeBackfill should respect lastReadSeqno monotonicity
The function already tried to avoid a 0 = 0 update (this happens in
unit tests). The MB shows that a stream with a non-zero start address
can have the same issue, but leads to for example 1 = 0.

Note cannot reproduce the issue seen in the MB, it appears to be a
view-engine stream which triggered the exception and logging is
disabled for those streams.

Change-Id: Ifc2565c58bd89820dadd2c7b621bf6c7a1787927
Reviewed-on: http://review.couchbase.org/124101
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>