MB-38356: Stream request for duplicate vbid stops streaming
If a DCP connection makes a stream request for a vbucket which it is
already streaming, then the second request should fail with
KEY_EEXISTS, and the initial stream remains as-is.

However, as part of changes for Collections to support multiple
StreamIDs for a single vBucket, this behaviour was regressed - the
KEY_EEXISTS is sent fine, however the initial stream incorrectly has
its cursor dropped (as the second essentially "steals" it). This
results in the initial stream hanging - no further mutations are sent
on it, even though it it still established.

Fix by splitting updateStreamsMap() into an initial check phase (which
will return the error /before/ creating a second ActiveStream) and a
subsequent update map phase.

Change-Id: I2887fabb6092e26da7838416a1edfb9f2c924686
Reviewed-on: http://review.couchbase.org/124117
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>