MB-39505: Add default stat label support to StatCollector
Prometheus supports labelled stats. These may share a metric name but
should have a unique set of name and labels.

As Prometheus will ingest stats from multiple buckets, all existing
stats will be labelled with the bucket name. Similar labelling will be
present for scopes and collections.

As these labels are common to groups of stats, and are not tied to a
_specific_ stat, they may be tracked by the StatCollector, rather than
being specified on a per-stat basis.

Support for per-stat labels will be added in later patches to allow
stats to be grouped into metric families but distinguished by labels.

Change-Id: I3ef48edfe2d3c8ca9e0ab5a4d25457aca5ff074f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128630
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>