MB-37921: Check for DcpStream in Dcp producer stream-request
Update the vbucket_filter so that privileges are checked.

The patch prepares the class for further privileges checks
by introducing a public checkPrivileges function. This
requires the internal set being changed to a map because
collection filter privilege checking needs scope and
collection ids and we post construction we cannot re-lookup
a scope-id in-case the manifest no longer knows of the
collection or scope.

Change-Id: I83c516d6f03cb340056510338d01eec12aa07313
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124537
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>