MB-38498: Don't return expired documents via GET_REPLICA
If a document is requested from a replica vBucket using the
GET_REPLICA command, don't return success if the document has actually
been expired.

Note this dates back to v3.0.0
(http://review.couchbase.org/c/ep-engine/+/30355) which changed the
logic to not expire replica items on access.

In the case of a replica, VBucket::fetchValidValue() returns early for
expired items if wantDeleted is true. However if not a replica then
the item is deleted before being returned. While `v` is returned in
both cases, in the case of the active is has now been deleted, but not
in the case of the replica (it is left as-is). In the calling function
(getInternal()) it then checks for v->isDeleted and() if true then
returns ENOENT.

Fix by checking if expired in getInternal(). This isn't ideal as we
are repeating the check, but returning null from
VBucket::fetchValidValue() fails in the full-eviction case as it will
get stuck in a loop returning EWOULDBLOCK to try to fetch an item from
disk.

Note this patch removes some expiration variants of rollback tests as:

1. They were previously broken - they were actually setting a TTL
   expiry to `expiryTime` seconds since unix epoch, not `expiryTime`
   from now. This meant they were already existed before any time
   passed.

2. Fixing them to use expiryTime-from-now highlighted a more
   fundamental flaw - after rollback, a TTL-expirated item _doesn't_
   become undeleted - the clock isn't rolled back so it doesn't make
   sense to expect them to be undeleted.

Change-Id: I9bc5a797bb250f7e2c40dad90d6490b44587b1d4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124515
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>