MB-38521: Don't allow concurrent execution of Producer closeAllStreams
Found a race condition during shutdown between the bucket destruction
thread and a worker thread with a DCP Producer.

The bucket destruction thread takes a copy of the cookieToConnection
map in the ConnMap and iterates over that copy calling closeAllStreams
on all connections. The worker thread calls disconnect on the ConnMap
(after the engine returns ENGINE_DISCONNECT to memcached) which calls
closeAllStreams. Both of these tasks are done outside of the
DcpConnMap::connsLock to prevent lock order inversions.

If both threads are allowed to concurrently execute
DcpProducer::closeAllStreams then we have a race condition where one
thread can destroy the backfillMgr before the other has finished
setting streams to dead. Settings streams to dead requires the
backfillMgr resulting in a segfault.

Fix this by taking a lock in DcpProducer::closeAllStreams to prevent
concurrent execution.

Change-Id: I7ba6dc5ee2d44fe112104a5692df07520600e532
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124795
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>