{Ephemeral,}VBucketTest: Fix failure when test run standalone
If these VBucket tests are run by themselves then they fail as
mock_server (required for TimeTraveller) has not been initialised.

The issue is that the test requires the use of the mock_server to
allow time to be advanced, however this test doesn't initialise the
mock time functions. This normally passes when run as part of the
entire ep-engine_ep_unit_test binary because an earlier test
initialises these time functions.

Change-Id: Ie4bf0d23ec338f3cf5a1d50573908cc6436bd8cc
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124845
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>