HashTableTest: Fix failure when test run standalone
If HashTableTest is run by itself
(e.g. --gtest_filter=*/HashTableTest.*") then testcase LockAfterDelete
fails:

    [ RUN      ] HashTableTest.LockAfterDelete
    ../kv_engine/engines/ep/tests/module_tests/hash_table_test.cc:994: Failure
    Expected: (osv->getCompletedOrDeletedTime()) >= (1985), actual: 0 vs 1985
    [  FAILED  ] HashTableTest.LockAfterDelete (1 ms)

The issue is that the test requires the use of the mock_server to
allow time to be advanced, however this test doesn't initialise the
mock time functions. This normally passes when run as part of the
entire ep-engine_ep_unit_test binary because an earlier test
initialises these time functions.

Change-Id: I51b15cd5dac4000c1e914bfee328fe6a3ae73745
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124818
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>