MB-38315: Evict deleted items in FullEviction
Currently we do not allow the eviction of deleted items. This means that
a BGFetch can load a deleted item into the HashTable which cannot be
removed by the item pager. Allow the eviction of deleted items from the
HashTable to free memory. This treats them in the same way as BGFetched
non-deleted items.

Change-Id: I00b8f8b8873021042373e29238f4b8077d8b7d30
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124832
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>