MB-38272: Share flusher batch limit by writers, not shards
Currently we are sharing the total flusher batch limit across all
shards. This is problematic for our default configuration though which
has 1 shard per core and 4 writer threads as it over-limits the amount
that a single thread can flush when we have more than 4 cores. It also
makes the flusher_total_batch_limit configuration parameter difficult
to use as it requires knowledge of both shards and writers to calcualte.

Split the flusher_total_batch_limit by writers instead of shards so that
we don't negatively affect the default configuration and make
calculating the per flush limit and max concurrent limit easier.

Change-Id: I4be7683dca3ee4dab9c972cd51d6bb7eeb8eb591
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124994
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>