MB-38272: Set flusher_total_batch_limit to pre-CheshireCat default
Before CheshireCat this value was set to 1,000,000 but this was a
per-thread limit. The default number of writer threads is 4 meaning that
by default we allow 4,000,000 items to be concurrently flushed. Now that
this value is divided by the number of writer threads we have, we should
multiply it before so that the default case is not affected. We still
want to limit cases where we have a very large number of writers to
try to cap memory usage whilst flushing.

Change-Id: Ic074fd278d64105bc0a98f91d8636442c7e0cdda
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124998
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>