MB-38634: Some configuration of SAN mapping won't work
If configured with:

    "client_cert_auth": {
        "state": "enable",
        "prefixes": [
          {
            "delimiter": ".",
            "path": "san.uri",
            "prefix": "www."
          },
          {
            "delimiter": "@",
            "path": "san.email",
            "prefix": ""
          }
        ]
    }

And we're providing a certificate without san.uri (but expecting a match
for san.email) the client used to be silently disconnected due an incorrect
error being returned from SanMapping::match.

This patch change the return value from Error to NoMatch which cause the
matcher to try the next rule instead of terminating.

Change-Id: I9cef908efd2141d96029b7b9127a2dc3bbe128bf
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/124999
Reviewed-by: Daniel Owen <owend@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>