MB-38528: Allow missing collection in setPersistedHighSeqno
Previously, when persisting a collection creation event, it was assumed
that the collection should exist when calling setPersistedHighSeqno.
However, a collection drop can remove the collection from the manifest
before the creation is persisted.

Make Collections::VB::Flush::setPersistedHighSeqno silently ignore
setPersistedHighSeqno for non-existent collections when persisting the
collection creation.

Change-Id: I54e95a7710f57982b4e03f3faabcf216625b9dd3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125051
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>