MB-38611: Magma - Set ConcurrentWriteCompact to Yes
Magma should allow writes concurrently with compactions. Currently
magma will run background compactions concurrently with writes; this
has nothing to do with the code in kv_engine. The ConcurrentWriteCompact
property prevents kv_engine from running compaction whilst flushing. Set
this property to Yes to allow compactions to run concurrently with
writes.

This change requires that we don't throw an exception if the compaction
can't find the vbstate as a unit test will fail otherwise. The root
cause of the unit test failing is actually that we are attempting to
compact a vBucket that no longer exists (we don't do this if
ConcurrentWriteCompact is set to No). To enable this test we should do
two things:

1: Only compact if the vBucket is still present when we start
   compaction.
2: Not throw if we can't find the vBucket, just return gracefully.
Change-Id: I93e613930259c593588fcfe887b9b363388dfbbb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125061
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>