MB-34694: Apply the Bucket Minimum Durability Level
With this change, the Durability Requirements of any write-request are
upgraded to the Bucket Minimum Durability Level, if the bucket min-level
is higher than the level specified by the user in the write-request.

The bucket min-level is applied at
EventuallyPersistentEngine::store/store_if/remove, as they handle all
the operations that we support for Durability.

Note that:
 - Applying the bucket min-level may turn a normal-write into a
   sync-write.
 - The level upgrade does not modify the original Durability Timeout,
   which stays unchanged.

Change-Id: I3f23c028b5976c956ca00d379af3f81a9444b3f3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125133
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>