MB-38656: Remove support from upgrading MetaData V0
In 3.0.0 we extended the metadata stored in the DocInfo to
include more information (datatype etc), and as part of
compaction we would inspect each document and replace
the content with a new DocInfo containing the new metadata
section.

Given that the only way to get into this situation would be
if someone took a CouchStore file written by a Couchbase
server < 3.0.0 (and never run compaction on the file) I think
it is safe to say that it is safe to kill the (complex) code.
We don't support any users running 2.x and if a user ever
want to go from 2.x to Cheshire-Cat they _MUST_ go through
a full compaction on one of the version in 3.x to 6.x
before doing the final upgrade to Cheshire-Cat.

The motivation for dropping this is that for every document
we inspect we try to parse the metadata section which
involves memory allocation. A database can have millions of
documents and run compaction over and over again and we'll
never end up seeing an object we want to upgrade.

If we ever end up in the situation where we want to reintroduce
the logic to rewrite the DocInfo structure we should keep
the status in a local document so we only try to inspect and
upgrade all documents one time (and not every time compaction
runs).

Change-Id: I5e17188fc14addb5f3aa11c297ec6db8a518c411
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125202
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>