MB-38673: Change PrivilegeAccess from enum to class
The intention behind this change is to allow for further detailed
failure states to be added in a second patch. If we had a new failure
values, code like ...

   if (state == PrivilegeAccess::Fail) {
       // Oh no failed
   } else {
       // cool we succeeded
   }

... would enter the wrong code.

Changing from an enum to a class allows for success()/failed() functions
to replace the above code so that when we do extend the failure states
such statements are executed.

   if (state.failed()) {
       // Oh no failed
   } else {
       // cool we succeeded
   }

Change-Id: I85137aa298d844210bbefc02fe27c809d2e8410a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125280
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>