MB-38673: Add new Privilege failure for no privileges
Add a new failure code to privilege checks which lets the caller
distinguish between a failure because the privilege wasn't found and a
failure where not only is the privilege not found but the user has no
privileges for the scope/collection.

The differing failure is important for returning an appropriate error
code based on the following:

1) When a user has no privileges for the scope or collection, it
   makes the scope or collection invisible. Status::FailNoPrivileges
2) When a user has some privileges for a scope or collection, but not
   the requested privilege the scope/collection is visible and the
   current operation can be failed with a no access error. Status::Fail

Internally to the privilege database the following logic is used to
return Status::Fail or Status::FailNoPrivileges.

The Bucket counts how many privileges in its set are collection
privileges (the bucket can have both collection privileges e.g. Read
and non collection e.g. Audit)
  - A collection privilege at the Bucket means every collection in the
    system gets that privilege, e.g. Read := you can Read everything.
  - When Bucket::check cannot find a requested collection privilege and
    no scope/collection search is possible it decides the error using
    the count of collection privileges - the rule being, >0 privileges
    makes collections visible.

The Scope::check now accepts a new parameter, which is the number of
collection privileges found in the parent (the Bucket)
  - The Scope can itself have an empty privilege set in the case where
    it is just a conduit from Bucket to Collection.
  - If the Scope::check fails it uses the count of its own privileges
    and the count of the parent privileges to determine Fail or
    FailNoPrivileges.

The Collection::check is unchanged - the assumption is that the
Collection privilege set should not be empty (now checked in
construction)
  - With the set always storing at least one collection privilege, if
    you reach Collection::check the only error can be Fail.

Change-Id: I2c8c1cc7bf27a132bb522839f70239d9cff3f029
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/125901
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>