MB-38497: Ensure verify_header() uses vbucket's collection manifest
Ensure McbpValidator::verify_header() uses the collection manifest at
the vbucket level. So can ensure that ops are valid against a
their vbucket's manifest as apposed to the global bucket manifest.
If this is not done we can end up in situations where DCP mutations for
a replica vbucket are rejected as the bucket level manifest hasn't been
update yet, but the vbucket's manifest has been.

Change-Id: Icb86a83f5a29077a4c83be05cb30a292dd70a220
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126504
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>