MB-37739: Make EPBucket::makeCompactionContext return shared_ptr
Currently EPBucket::makeCompactionContext returns a compaction_ctx
object. This doesn't suit the magma compaction API well though which
re-uses the same struct for both explicit and implicit compactions. This
is because explicit compactions require EPBucket to own the
compaction_ctx and implicit compactions require magma (or rather the
magma compaction callback object) to own the compaction_ctx. To allow
a single struct  to cater for both of these cases, make
EPBucket::makeCompactionContext return a shared_ptr so that a single
field can be used for the compaction_ctx in the magma compaction
callback.

Change-Id: I6e63d91c38319e96d60957806a51b32a13a44a7d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126018
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Scott Lashley <scott.lashley@couchbase.com>