MB-39046: Introduce units for statistics
While many existing stats represent a value with units (e.g.,
microseconds), this was not always programatically determinable.

In order to correctly expose stats for Prometheus, stats need to be
scaled to base units (microseconds->seconds, kilobytes->bytes) and
appropriately suffixed ( https://prometheus.io/docs/practices/naming/
).

This patch introduces Unit types which can be used to normalise values
and determine the correct suffix.

Units are not yet used in this patch, but will in future be used in
metric definitions.

Change-Id: I656f13406e8b039e54c84bdb464c45012ee92c1f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128633
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>