MB-33463: Add support for magma GetDocs API
The Magma::GetDocs API allows for the submission of multiple Get
requests where those Get requests are broken up across coroutines in
magma to increase parallelism. Because the list submitted to getMulti is
a map rather than a vector or deque, we need to transfer the list into a
Magma::OperationsList.

Change-Id: Iff70f5b8a40f00b2cf824152d0216207e2ed7ca0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126160
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>