[cluster_test] Don't throw exceptions from destructor
As part of shutdown cluster_test tries to clean up the
temporary directories and configuration files used while
running the test. If there was a problem deleting a file
cb::io::rmrf throws an exception, and if this happens
in the destructor the runtime on Windows will terminate
the program by calling abort() (which ends up with exit(3)
being called making it hard to figure out why the program
silently failed).

Change-Id: Ief0abd648e4b8a1b7a07d1de528e3706177d0ac7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126480
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>