[cluster_test] Retry file removal
Cluster_test creates a new directory for each cluster being
started, where each node get its own directory for database
and logging. During shutdown we try to remove these files
to avoid filling up the disk space with old and obsolete files.

When running on windows we see sporadic failures trying to
delete these files (EINVAL being reported back). I believe
this is a mapping of another error code, so I tried to use
the DeleteFile from the windows API instead, and it returned
EIO which could indicate that someone else have the file
open.

During shutdown we do try to stop all of the nodes by
clearing the vector of nodes (which should stop the processes
and release all resources _BEFORE_ we start deleting the
files), so it is not fully clear to me _WHY_ we're getting
this return code from remove (it could be that the process
isn't fully cleaned up within the kernel even if we read
out the exit status with GetExitCodeProcess, but the
kernel is currently running the cleanup "in parallel").
Anyway, it turns out that just backing off and retry
solves the problem.

Before this patch we would "leak" one directory relatively
often (from cluster_test --gtest_repeat=1000 I saw the
number of directories slowly grow). With this patch I've
not seen any directories leaked (note: the current patch
gives up deleting the directory after 10 seconds)

Change-Id: I6d5b7b587d77f8c13b4e5c2bf03e329583d4a7c4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126492
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>