MB-38856: Handle dropping sync-writes from ephemeral buckets
The previous code that dropped a key was not handling the case
where a pending and committed item both exist (an ephemeral only
case). The code was leading to a pointer to freed memory being left
in the hash-table leading to various crashes.

Note that this MB does not fully address or test sync-writes and
collection drops which is tracked under MB-34217.

Change-Id: I21f2764abdaa37c2eddf4f32999a64321cc3ae69
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126619
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>