MB-38849: Remove correct # items in setUpSendSetInsteadOfCommitTest
Depending on how slow/fast flushing items is, the way checkpoint
boundaries fall can differ in unit tests. In
setUpSendSetInsteadOfCommitTest we assert that we remove 4 items from
the checkpoint manager. To ensure that we always remove 4 items, create
a new checkpoint before we set/flush the final item. This ensures that
only 1 item will exist in the last checkpoint in the CheckpointManager
when we call removeClosedUnrefCheckpoints which in turn ensures that we
can always remove 4 items (all the others in the checkpoint manager).

Change-Id: I4c26918230195cd737df089dc6e929258ff74d61
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126813
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>