MB-39027: Fix completeBackfill() scan range over openChk range
Fix ActiveStream::completeBackfill() to use last read seqno from the
KVStores's scan context when making a decision of whether we
should consider enqueuing a SeqnoAdvanced op. As this allows us to
ensure when we're a replica and performing back fill that we have
indeed come to the end of the snapshot or if we're about to
transition streaming the rest of the snapshot from memory. In which
case we do not want to consider enqueuing a SeqnoAdvanced as we are
in the middle of a snapshot. It is worth noting this only effects
replica vbuckets.

Change-Id: I2091e8acd4a62cc83ce96f67d3144512c3cb41c5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126851
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>