Make deduction of getBucketMin/Max template arguments possible
Dependent types don't allow template arguments be deduced from
parameters.

This means that

 Histogram<T, Limits> hist;
 getBucketMin(*hist.begin());

would not compile, and would instead need to be

 getBucketMin<T, Limits>(*hist.begin());

By chance, all remaining uses of Histogram which were formatted as
stats happen to be

 using MicrosecondHistogram =
         Histogram<UnsignedMicroseconds, cb::duration_limits>;

Which has a separate overload, so the templated version was not
instantiated until additional tests were added in a later patch.

This patch avoids using a dependent type in the function parameter.

 getBucketMin(...);

can now be used as intended from templated code.

Change-Id: I9a291be0792a58cb164882f697bf152a98906b69
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/126841
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>