MB-39113: Ensure we don't crash if we purge a lower seqno seen before
Ensure we don't crash due to a monotonic exception being thrown. This
exception is caused when we self assign highestPurgedDeletedSeqno to
its self during BasicLinkedList::purgeListElem(), when we purge an item
which has a seqno less than its current value. This is a valid situation
that may occur when items with lower seqno are marked as deleted between
purges. To fix this ensure we only update the value of
highestPurgedDeletedSeqno when the item we're purging is greater than
its current value.

Change-Id: I3fa760dba42fa06a617856f330a609e4729fa232
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127508
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>