MB-38845: Do not update stats at flush-failure for VBs set for deletion
For some reason, we used to update some flush stats in
PersistenceCallback::redirty(), which is wrong as that function is
executed in the code path for flush-failure.
That may cause underflow in some stats. In particular
VBucket::dirtyQueueSize is a NonNegativeCounter and may throw in
dev-builds.

Change-Id: I72931613422205139f71c359b9305e24abe1b465
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127190
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>