MB-38845: Remove PersistenceCallback::redirty
In the callback for set:
  1) first, we mark a StoredValue clean
  2) then, we check is flush has failed and we reset it to dirty
Step (2) happens in redirty().

With this patch we just mark a StoredValue clean if the flush is
success, so redirty() becomes pointless.

Also, the call was already pointless for callback::del(), given that a
delete exists dirty in the HastTable or it doesn't exist at all (as it
is removed from the HT at persistence).

Change-Id: I7f60608cacffdd0c7052b3434e6b46656902ee2a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127242
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>