MB-37319: Allocate enough space for extended Snapshot Marker
When encoding the DCP Snapshot Marker we allocate the space
before using the FrameBuilder to encode the frame to put on
the wire. With the introduction of version 2.1 we add an extra
field to the packet, so we need to allocate space for it.

This is missing from 4459fdca5188642867f55ba999f31d8e6e60174c
which added the timestamp to the packet, but we currently
don't try to encode or send such packets in our system
yet.

Change-Id: Ie7a843f9de4bd253b81a4dcb2ae7721b5d6c8d9f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127306
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>