MB-37319: Add parameter to initBySeqnoScanContext
This is a preparation to add support for PointInTime capable
DCP streams by allowing initBySeqnoScanContext to initialize
the scan context from a "historical" snapshot instead of a
current snapshot.

This patch adds the parameter (no new functionality) and make
sure that we call the method with the value to provide the
same behaviour as we used to have.

Change-Id: I11424ad2c8478591d3b482223ac59865d7208df0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127318
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>