MB-37319: Add PiTR knowledge to ActiveStream
This patch populates the knowledge of the PiTR enabled feature
for the DCP producers down to the ActiveStream class and initialize
the Scan Context requesting a historical snapshot if enabled.

Currently the underlying kvstore "ignores" this and gives the
current data (as the code to open a historical snapshot is moved
to a separate patch to minimize the changeset and ease the review).

Change-Id: I4eeb68b4952e0177dd4942441aa3d8d0d39800f7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127319
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>