MB-16181: Use UpgradeHolder in collections manifest update
Make use of the UpgradeHolder so that we can reduce the scope of
the write locking when doing a VB::Manifest update.

The first part of the update is validation and figuring out the set
of changes to make, this only needs read locking on the VB::Manifest.

Once validation is complete and the set of changes to make is known
we need write locking to make the changes.

The UpgradeHolder is designed to facilitate this allowing an atomic
switch from read (upgrade) to write.

Change-Id: Ie803d54683a28e4f2e8ef4648ad855527d8e4582
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127518
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>