MB-37680: BackfillManager::backfill() - Simplify snooze logic
In BackfillManager::backfill() if the DCPBackfill returns 'snooze',
then it is put onto the snoozingBackfills list for later
scheduling. However, before putting it on the list we check if that
VBucket is still valid; if not the Backfill is cancelled.

This is essentially pointless - while the VBucket may be valid for
that instant, one moment after we have added the DCPBackfill to
snoozingBackfills it could be deleted or similar - and the backfill
has already been snoozed. As such, remove this code to simply
BackfillManager.

Note that each Backfill class checks for Stream validity every time
the backfill runs, which ensures the backfill is cancelled (finished)
if the VBucket is no longer valid.

Change-Id: Ibe6e66571299815f2cf21ccb97c583ec60a98acb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/127873
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>