MB-35889: Don't add keys to Checkpoint indexes for Disk checkpoints
Disk checkpoint indexes grow at a O(n) rate which is unsustainable for
heavy DGM use cases as we can easily run out of memory. We could expel
the keys from these indexes along with Checkpoint expelling and keep
most of our SyncWrite sanity checks that use the checkpoint key index.
Alternatively, we can just not add the keys to the indexes for Disk
checkpoints. This means that we won't perform any of the SyncWrite
sanity checks for Disk checkpoints.

Given that we don't expect a Disk checkpoint to ever contain the same
key more than once and our desire to reduce memory overhead/maintain a
good working set, don't add the keys to the checkpoint indexes for Disk
checkpoints.

Change-Id: Id27dd7ebe8edc318e9b0108499f494d542b0e8ae
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128087
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>